refactor(mocks): extract forge module from mockApi (#991 M3)
Move connections/repos/github-project-sync/tool-allowlist state and the ~36 forgeConfig…setRepoToolProfile methods plus adminListBlockedRepos into forge.ts as forgeApi; index.ts spreads it and imports repos for the findings/schedules/admin reads that share it.
